[UPLIFT-Configurator] Added a configurator for BikeShareViewController
