+++ Podfile
@@ -1,15 +1,15 @@
 # Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
+platform :ios, ’10.0’
 
 target 'BikeShare' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
   use_frameworks!
 
   # Pods for BikeShare
- pod 'Alamofire', '~> 4.7'
+  pod 'Alamofire', '~> 4.7'
   pod 'GoogleMaps'
   pod 'AlamofireObjectMapper', '~> 5.0'
-
+  
   target 'BikeShareTests' do
     inherit! :search_paths
     # Pods for testing
